Moved ship stuff into a subdirectory
