+++ src/rootReducer.test.js
@@ -31,13 +31,11 @@ test('showWelcomePage off and on', () => {
 
 test('Buy a ship', () => {
     let s0 = state0()
-    let expectedShip = {shipId: 'ship1', shipType: 'fishboat', shipName: 'The Flying Flea', basePrice:500, shipTypeName:'fishboat'}
     expect(actions.buyShip('testid')).toEqual({type: 'BUY_SHIP', shipId: 'testid'})
     let s1 = rootReducer(s0, actions.buyShip('ship1'))
-    console.log('s1='+JSON.stringify(s1))
     expect(s1.myShipId).toBe('ship1')
     expect(s1.cash).toBe(300)
-    expect(stateUtils.getMyShip(s1)).toEqual(expectedShip)
+    expect(stateUtils.getMyShip(s1).shipId).toBe('ship1')
     expect(stateUtils.getDynamicPlaceById('portharbor', s1).shipsForSale.indexOf('ship1')).toBe(-1)
 });
