+++ src/ship/Cash.js
