+++ src/ship/ShiplessPanel.js
